+++ crates/settings/src/settings_content/editor.rs
@@ -95,6 +95,10 @@ pub struct EditorSettingsContent {
     /// Default: 4.0
     #[serde(serialize_with = "crate::serialize_optional_f32_with_two_decimal_places")]
     pub fast_scroll_sensitivity: Option<f32>,
+    /// Whether to animate scrolling with a smooth easing effect.
+    ///
+    /// Default: disabled
+    pub smooth_scroll: Option<SmoothScrollContent>,
     /// Settings for sticking scopes to the top of the editor.
     ///
     /// Default: sticky scroll is disabled
@@ -318,6 +322,20 @@ pub struct StickyScrollContent {
     ///
     /// Default: false
     pub enabled: Option<bool>,
+}
+
+/// Smooth scroll related settings
+#[with_fallible_options]
+#[derive(Clone, Default, Debug, Serialize, Deserialize, JsonSchema, MergeFrom, PartialEq)]
+pub struct SmoothScrollContent {
+    /// Whether smooth scroll is enabled.
+    ///
+    /// Default: false
+    pub enabled: Option<bool>,
+    /// Scroll animation duration.
+    ///
+    /// Default: 125
+    pub duration: Option<DelayMs>,
 }
 
 /// Minimap related settings